+++ src/client/editor/index.js
@@ -209,6 +209,7 @@ class Editor {
 
             for (let c of combos) {
                 keyboardJS.bind(c, (e)=>{
+                    e.catchedByEditor = true
                     this.handleKeyboard(c, e)
                 })
             }
@@ -330,10 +331,10 @@ class Editor {
 
                 break
 
-            case 'up':
-            case 'down':
-            case 'right':
-            case 'left':
+            case 'mod + up':
+            case 'mod + down':
+            case 'mod + right':
+            case 'mod + left':
                 if (!this.selectedWidgets.length) return
 
                 if (this.selectedWidgets[0].props.top === undefined && e.key.match(/Arrow(Up|Down)/)) return
@@ -368,33 +369,24 @@ class Editor {
                 }
                 break
 
-            case 'mod + up':
-            case 'mod + down':
-            case 'mod + right':
-            case 'mod + left':
+            case 'up':
+            case 'down':
+            case 'right':
+            case 'left':
                 if (!this.selectedWidgets.length) return
 
                 var curWidget = this.selectedWidgets[0],
                     toSelect = null
 
-                if (e.key === 'ArrowUp' && curWidget.parent !== widgetManager) {
+                if (e.key === 'ArrowLeft' && curWidget.parent !== widgetManager) {
 
                     toSelect = curWidget.parent
 
-                } else if (e.key === 'ArrowDown') {
+                } else if (e.key === 'ArrowRight') {
 
-                    let toSelectList = [...curWidget.children]
+                    toSelect = curWidget.children[0]
 
-                    if (toSelectList && toSelectList.length) {
-                        toSelectList.sort((a,b) => a.container.offsetLeft>b.container.offsetLeft)
-                        toSelect = toSelectList[0]
-                        if (toSelect.container.classList.contains('not-editable')) {
-                            toSelect = null
-                        }
-                    }
-
-                }
-                else if((e.key == 'ArrowLeft') || (e.key == 'ArrowRight')){
+                } else if((e.key == 'ArrowUp') || (e.key == 'ArrowDown')){
 
                     if (curWidget.parent === widgetManager) return
 
@@ -404,7 +396,7 @@ class Editor {
                         toSelectList.sort((a,b) => a.container.offsetLeft > b.container.offsetLeft)
                         var idx = toSelectList.indexOf(curWidget)
                         if (idx >= 0) {
-                            var nextIdx = (idx + (e.key === 'ArrowLeft' ? -1 : 1 ) + toSelectList.length) % toSelectList.length
+                            var nextIdx = (idx + (e.key === 'ArrowUp' ? -1 : 1 ) + toSelectList.length) % toSelectList.length
                             toSelect = toSelectList[nextIdx]
                         }
                     }
@@ -828,12 +820,16 @@ class Editor {
         for (var i = 0; i < this.selectedWidgets.length; i++) {
 
             let w = this.selectedWidgets[i],
-                dW = deltaW * w.container.offsetWidth / this.widgetDragResize.initWidth,
-                dH = deltaH * w.container.offsetHeight / this.widgetDragResize.initHeight,
                 nW, nH
 
-            nW = w.container.offsetWidth + dW
-            nH = w.container.offsetHeight + dH
+            if (i === 0 && ui) {
+                nW = ui.originalSize.width + deltaW
+                nH = ui.originalSize.height + deltaH
+
+            } else {
+                nW = w.container.offsetWidth + deltaW
+                nH = w.container.offsetHeight + deltaH
+            }
 
             if (w.props.width !== undefined && w.parent.getProp('layout') !== 'vertical') {
                 var newWidth = Math.max(nW, GRIDWIDTH) / PXSCALE