+++ src/client/widgets/containers/matrix.js
@@ -17,12 +17,12 @@ class Matrix extends Panel {
             class_specific: {
                 widgetType: {type: 'string', value: 'button', help: 'Defines the type of the widgets in the matrix'},
                 quantity: {type: 'number', value: 4, help: 'Defines the number of widgets in the matrix'},
-                start: {type: 'integer', value: 0, help: 'First widget\'s index'},
                 props: {type: 'object', value: {}, help: [
                     'Defines a set of property to override the widgets\' defaults.',
-                    'JS{{}} and #{} blocks in this field are resolved with an extra variable representing each widget\'s index: `$` (e.g. `#{$}`)',
-                    'Advanced syntax blocks (@{}, OSC{}, JS{{}} and #{}) are resolved at the matrix\' scope (ie @{this.variables} returns the matrix\' variables property)',
-                    'Note: by default, the widgets inherit from the matrix\' `id` and osc properties (`id` and `address` are appended with `/$`)'
+                    'JS{} and #{} blocks in this field are resolved with an extra variable representing each widget\'s index: `$` (e.g. `#{$}`)',
+                    'Advanced syntax blocks (@{}, OSC{}, JS{}, VAR{} and #{}) are resolved at the matrix\' scope (ie @{this.variables} returns the matrix\' variables property)',
+                    'Advanced syntax blocks can be passed to children without being resolved at the matrix\' scope by adding an underscore before the opening bracket.',
+                    'Note: unless overridden, children inherit from the matrix\' `id` and osc properties (`id` and `address` are appended with `/$`)'
                 ]},
             },
             style: {
@@ -60,18 +60,18 @@ class Matrix extends Panel {
 
 
 
-        this.start = parseInt(this.getProp('start'))
-
         if (parser.widgets[this.getProp('widgetType')]) {
 
             for (let i = 0; i < this.getProp('quantity'); i++) {
 
-                var props = this.resolveProp('props', undefined, false, false, false, {'$':i + this.start})
-                var data = this.defaultProps(i + this.start)
+                var props = this.resolveProp('props', undefined, false, false, false, {'$':i})
+                var data = this.defaultProps(i)
 
                 if (typeof props === 'object' && props !== null) {
                     Object.assign(data, props)
                 }
+
+                data = JSON.parse(JSON.stringify(data).replace(/(JS|#|OSC|@|VAR)_\{/g, '$1{'))
 
                 var widget = parser.parse({
                     data: data,
@@ -126,11 +126,13 @@ class Matrix extends Panel {
 
                 for (let i = this.children.length - 1; i >= 0; i--) {
 
-                    let data = this.resolveProp('props', undefined, false, false, false, {'$': i + this.start})
+                    let data = this.resolveProp('props', undefined, false, false, false, {'$': i})
 
                     if (typeof data !== 'object' || data === null || Array.isArray(data)) {
                         data = {}
                     }
+
+                    data = JSON.parse(JSON.stringify(data).replace(/(JS|#|OSC|@|VAR)_\{/g, '$1{'))
 
                     if (typeof oldPropValue === 'object' && oldPropValue !== null) {
                         // if an overridden default props is removed, set it back to default
@@ -139,7 +141,7 @@ class Matrix extends Panel {
                         for (var k in oldPropValue) {
                             if (data[k] === undefined) {
                                 if (Matrix.overriddenDefaults.indexOf(k) !== -1) {
-                                    overriddenDefaults = overriddenDefaults || this.defaultProps(i + this.start)
+                                    overriddenDefaults = overriddenDefaults || this.defaultProps(i)
                                     data[k] = overriddenDefaults[k]
                                 } else {
                                     widgetDefaults = widgetDefaults || parser.defaults[data.type || this.getProp('widgetType')]